MainView styles in their own file
